refactors components in `Settings` Page into separate files to contain individual settings logic in it's own file
