+++ src/ui/components/settings/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2022 @paritytech/contracts-ui authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export * from './CustomEndpoint';
+export * from './ThemeMode';
